LPD-30131 Refactor to use helpers
